몬스터 출력
. 서버에서 제공되고 있던 몬스터 관련 패킷을 받아 클라이언트에서 출력한다.
. 몬스터 애니메이션은 적용되지 않은 상태.

+++ 소스/WorldEnd/Client/mesh.h
@@ -173,7 +173,7 @@ public:
 	vector<string>& GetSkinningBoneNames() { return m_skinningBoneNames; }
 	vector<shared_ptr<GameObject>>& GetSkinningBoneFrames() { return m_skinningBoneFrames; }
 
-	UINT GetSkinningBoneNum() const { return m_skinningBoneFrames.size(); }
+	UINT GetSkinningBoneNum() const { return (UINT)m_skinningBoneFrames.size(); }
 	string GetSkinnedMeshName() const { return m_skinnedMeshName; }
 private:
 	string								m_skinnedMeshName;
